Add order authorization trait
